+++ changelog.md
@@ -1,4 +1,8 @@
 # ChangeLog
-- 修复在模块禁用时，开机模块状态显示错误
-- 优化了模块状态显示emoji
-- 优化了默认模块配置
+- 新增 action.sh 用于开关模块
+- 新增 webroot 用于管理 AdGuardHome
+- 微调了 AdGuardHome.yaml 更新了秋风广告规则
+- 删除 agh_service.sh 转而使用 service.sh
+- 更完全地禁用ipv6
+- 删除了安装脚步中是否禁用iptables的选项
+- 安装脚本现在会根据系统语言本地化输出